qreal should not be used in dbus interfaces, as otherwise in platforms where qreal = float functions with qreal's won't be seen at all thus, just use double and in case cast to float
svn path=/branches/KDE/4.2/kdebase/workspace/; revision=923562

+++ libs/kephal/kded_kephal/dbus/dbusapi_configurations.cpp
@@ -109,10 +109,10 @@ bool DBusAPIConfigurations::rotate(QString output, int rotation) {
     return false;
 }
 
-bool DBusAPIConfigurations::changeRate(QString output, qreal rate) {
+bool DBusAPIConfigurations::changeRate(QString output, double rate) {
     Output * o = Outputs::self()->output(output);
     if (o) {
-        return Configurations::self()->changeRate(o, rate);
+        return Configurations::self()->changeRate(o, static_cast<float>(rate));
     }
     return false;
 }
